Backend/mitapi/requirements.txt
APScheduler==3.11.0
asgiref==3.9.1
azure-core==1.35.0
azure-storage-blob==12.25.1
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
defusedxml==0.7.1
diff-match-patch==20241021
Django==5.2.4
django-anymail==13.0
django-cors-headers==4.7.0
django-filter==25.1
django-import-export==4.3.8
django-simple-history==3.10.1
django-upgrade==1.25.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
gunicorn==23.0.0
idna==3.10
isodate==0.7.2
MarkupPy==1.18
#odfpy==1.4.1
openpyxl==3.1.5
packaging==25.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
pyotp==2.9.0
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.2
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
tokenize_rt==6.2.0
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
# uWSGI==2.0.30
xlrd==2.0.2
xlwt==1.3.0
